hwif_wrapper_tool/pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "peakrdl-hwif-wrapper"
version = "0.1.0"
description = "Standalone tool to generate hwif wrapper for PeakRDL-regblock"
requires-python = ">=3.8"
dependencies = [
    "systemrdl-compiler>=1.27",
    "peakrdl-regblock>=1.0",
    "Jinja2>=2.11",
]

[project.scripts]
peakrdl-hwif-wrapper = "hwif_wrapper_tool.cli:main"

